pyproject.toml
[tool.poetry]
name = "population_analysis"
version = "0.1.0"
description = "Analysis of the outputs from the MOBSE and FASCLUSTER codes."
authors = ["jacopok <jacopok@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
pre-commit = "^2.17.0"
scipy = "^1.8.0"
imageio = "^2.16.1"
ipython = "^8.4.0"
pandas = "^1.4.2"
tqdm = "^4.64.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
